Using Jansi instead of Google Lantern. Using the same logic for running a single test.
