+++ conductor
@@ -2,8 +2,5 @@
 
 set -e
 
-if [ $# -eq 0 ]; then
-  ./gradlew :conductor-cli:run -q --console=plain
-else
-  ./gradlew :conductor-cli:run -q --console=plain --args="$*"
-fi
+./gradlew :conductor-cli:shadowJar
+java -jar conductor-cli/build/libs/conductor-cli-*-all.jar "$@"
