+++ conductor-cli/src/main/java/conductor/cli/command/TestCommand.kt
@@ -1,14 +1,10 @@
 package conductor.cli.command
 
-import conductor.Conductor
-import conductor.ConductorException
-import conductor.cli.continuous.ContinuousTestRunner
+import conductor.cli.runner.ContinuousTestRunner
+import conductor.cli.runner.SingleTestRunner
 import conductor.cli.util.ConductorFactory
 import conductor.orchestra.CommandReader
-import conductor.orchestra.ConductorCommand
-import conductor.orchestra.Orchestra
 import conductor.orchestra.yaml.YamlCommandReader
-import okio.source
 import picocli.CommandLine
 import picocli.CommandLine.Option
 import java.io.File
@@ -47,41 +43,13 @@ class TestCommand : Callable<Int> {
         }
 
         val conductor = ConductorFactory.createConductor(target)
+        val commandReader = resolveCommandReader(testFile)
 
         return if (continuous) {
-            ContinuousTestRunner(conductor, testFile).run()
+            ContinuousTestRunner(conductor, testFile, commandReader).run()
             0
         } else {
-            runSingleTest(conductor)
-        }
-    }
-
-    private fun runSingleTest(conductor: Conductor): Int {
-        val commands = readCommands(testFile)
-
-        conductor.use {
-            println("Running test on ${it.deviceName()}")
-
-            try {
-                Orchestra(it).executeCommands(commands)
-            } catch (e: ConductorException) {
-                println(e.message)
-
-                println("FAILURE")
-                return 1
-            }
-        }
-
-        println("SUCCESS")
-
-        return 0
-    }
-
-    private fun readCommands(file: File): List<ConductorCommand> {
-        val reader = resolveCommandReader(file)
-
-        return file.source().use {
-            reader.readCommands(file.source())
+            SingleTestRunner(conductor, testFile, commandReader).run()
         }
     }
