+++ conductor-cli/src/main/java/conductor/cli/runner/CommandState.kt
@@ -1,0 +1,8 @@
+package conductor.cli.runner
+
+import conductor.orchestra.ConductorCommand
+
+data class CommandState(
+    val status: CommandStatus,
+    val command: ConductorCommand,
+)
