+++ conductor-cli/src/main/java/conductor/cli/runner/CommandStatus.kt
@@ -1,0 +1,8 @@
+package conductor.cli.runner
+
+enum class CommandStatus {
+    PENDING,
+    RUNNING,
+    COMPLETED,
+    FAILED,
+}
